requirements.txt
Flask==3.0.3
docker==7.1.0
pymongo==4.9.1
psutil==6.0.0
python-dotenv==1.0.1

